requirements.txt
aiogram==2.25.1
python-dotenv
SQLAlchemy
asyncpg
openai
stripe
aiohttp
psycopg2-binary
